+++ src/main/java/it/consciousdreams/ActionEditDialog.java
@@ -17,9 +17,9 @@ import java.awt.*;
 import java.awt.event.KeyAdapter;
 import java.awt.event.KeyEvent;
 import java.io.File;
-import java.util.List;
 
 import static java.awt.event.KeyEvent.*;
+import java.util.List;
 
 public class ActionEditDialog extends DialogWrapper {